requirements.txt
fastai==2.5.3
fastbook==0.0.18
Flask==2.0.2
webdriver_manager==3.5.2
Werkzeug==2.0.1
Pillow